Moved the following projects to NHibernate Contrib (http://sourceforge.net/projects/nhcontrib/) - NHibernate.Search
Note: The exported revision is of NHibernate.Search is r3325. There had been some breaking changes in Core between r3325 and r3343. Dario Quintana kindly fixed that for NH.Search. Unfortunately those changes cannot be moved to nhcontrib as the NHibernate.dll in nhcontrib has not been updated to include those breaking changes. When NHibernate.dll is updated in nhcontrib, a diff can be made and Dario's changes can be re-incorporated into NH.Search in nhcontrib.

git-svn-id: https://nhibernate.svn.sourceforge.net/svnroot/nhibernate/trunk/nhibernate@3344 d2eaab8a-a80d-0410-be94-99ecdb4ea5df